Split benchmarkers into different files
