+++ benchmark.go
@@ -25,6 +25,7 @@ import (
 	"errors"
 	"fmt"
 	"math"
+	"math/rand"
 	"os"
 	"os/signal"
 	"runtime"
@@ -32,6 +33,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/opentracing/opentracing-go"
 	"github.com/yarpc/yab/limiter"
 	"github.com/yarpc/yab/statsd"
 	"github.com/yarpc/yab/transport"
@@ -47,6 +49,17 @@ var (
 	// the same array around to multiple functions
 	_quantiles = []float64{0.5000, 0.9000, 0.9500, 0.9900, 0.9990, 0.9995, 1.0000}
 )
+
+type peerTransport struct {
+	transport.Transport
+	peerID int
+}
+
+type benchmarker interface {
+	WarmTransports(int, TransportOptions, resolvedProtocolEncoding, int) ([]peerTransport, error)
+	Call(transport.Transport) (time.Duration, error)
+	Method() string
+}
 
 // Parameters holds values of all benchmark parameters
 type Parameters struct {
@@ -311,4 +324,65 @@ func stopOnInterrupt(out output, r *limiter.Run) {
 		out.Printf("\n!!Benchmark interrupted!!\n")
 		r.Stop()
 	}()
+}
+
+func peerBalancer(peers []string) func(i int) (string, int) {
+	numPeers := len(peers)
+	startOffset := rand.Intn(numPeers)
+	return func(i int) (string, int) {
+		offset := (startOffset + i) % numPeers
+		return peers[offset], offset
+	}
+}
+
+// warmTransport warms up a transport and returns it. The transport is warmed
+// up by making some number of requests through it.
+func warmTransport(m benchmarker, opts TransportOptions, resolved resolvedProtocolEncoding, warmupRequests int) (transport.Transport, error) {
+	transport, err := getTransport(opts, resolved, opentracing.NoopTracer{})
+	if err != nil {
+		return nil, err
+	}
+
+	for i := 0; i < warmupRequests; i++ {
+		_, err := m.Call(transport)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return transport, nil
+}
+
+// warmTransports returns n transports that have been warmed up.
+// No requests may fail during the warmup period.
+func warmTransports(m benchmarker, n int, tOpts TransportOptions, resolved resolvedProtocolEncoding, warmupRequests int) ([]peerTransport, error) {
+	peerFor := peerBalancer(tOpts.Peers)
+	transports := make([]peerTransport, n)
+	errs := make([]error, n)
+
+	var wg sync.WaitGroup
+	for i := range transports {
+		wg.Add(1)
+		go func(i int, tOpts TransportOptions) {
+			defer wg.Done()
+
+			peerHostPort, peerIndex := peerFor(i)
+			tOpts.Peers = []string{peerHostPort}
+
+			tp, err := warmTransport(m, tOpts, resolved, warmupRequests)
+			transports[i] = peerTransport{tp, peerIndex}
+			errs[i] = err
+		}(i, tOpts)
+	}
+
+	wg.Wait()
+
+	// If we hit any errors, return the first one.
+	for _, err := range errs {
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return transports, nil
 }
